% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{.match_reference_component}
\alias{.match_reference_component}
\title{Match Reference Component}
\usage{
.match_reference_component(reference, X)
}
\arguments{
\item{reference}{Column index or column name.}

\item{X}{Numeric matrix.}
}
\value{
Integer column index.
}
\description{
Internal helper for reference-based coordinates.
}
\keyword{internal}
